requirements.txt
fastapi>=0.115.12
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
xata
python-dotenv
instructor
openai
weaviate-client